ref: ether api. refs: #21
